Move CompactMerkleTree to a separate package
