+++ merkle/compact/compact_merkle_tree.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package merkle
+package compact
 
 import (
 	"bytes"
